refactor(cli): move run wait to run_wait.go (PRD #1009 M1)
